Release keys when emulator window loses focus.
Otherwise keys get "stuck down", since we don't see key-release events
when the window doesn't have input focus.  In particular, if pressing
a key causes the calc to hit a breakpoint, the key-release event will
be sent to the debugger window instead of the emulator.

Treating all keyboard keys as "released" when we lose focus is
probably good policy anyway.


git-svn-id: https://svn.code.sf.net/p/tilem/code/trunk@716 05cf5367-e1a8-405a-8a5c-7f956f11530a