Fix copying of partial audio data buffers.
Previous code was copying the wrong part of the buffer (thereby
producing garbled output) if the driver used a smaller buffer size
than expected (or, I guess, a larger buffer size that isn't a power of
two.)  This happens to occur with the pulseaudio driver, as well as
the with the waveout driver on Windows.  It may be the case for other
drivers as well.


git-svn-id: https://svn.code.sf.net/p/tilem/code/trunk@739 05cf5367-e1a8-405a-8a5c-7f956f11530a