Fix subdirectory compilation with certain versions of 'make'.
Some versions of 'make' will run all commands for a given rule with a
single invocation of the shell, rather than treating each line as a
separate command.  Apparently, BSD make will do so only if the -j
option is used.  As a result, commands that 'cd' to a subdirectory may
or may not affect the working directory for subsequent commands.


git-svn-id: https://svn.code.sf.net/p/tilem/code/trunk@740 05cf5367-e1a8-405a-8a5c-7f956f11530a