Better emulation of ILI9335 pixel read operations.
Experiments indicate that if a pixel read operation is interrupted
(i.e., a program reads the high byte of a pixel, then changes the IR
or writes a byte), the low byte of the input pixel is discarded.  So
if the program continues reading pixels afterwards, it will begin with
the high byte of the next pixel.

Furthermore, it appears that the next pixel is loaded from GRAM at the
time of the first IN operation (so at that point, there are three
bytes "queued" in internal registers.)

(Warning: this change breaks compatibility with existing 84+CSE state
files.  It should not have any visible effect unless you have saved
the calculator state in the middle of a subroutine that reads LCD
pixel data.)


git-svn-id: https://svn.code.sf.net/p/tilem/code/trunk@750 05cf5367-e1a8-405a-8a5c-7f956f11530a